Agrego información de aplication properties

+++ src/main/java/com/turings/backend/model/Categoria.java
@@ -7,7 +7,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private  Long id_categoria;
+    private  int id_categoria;
 
     private String categoria;
 
@@ -17,11 +17,11 @@ public class Categoria {
         this.categoria = categoria;
     }
 
-    public Long getId_categoria() {
+    public int getId_categoria() {
         return id_categoria;
     }
 
-    public void setId_categoria(Long id_categoria) {
+    public void setId_categoria(int id_categoria) {
         this.id_categoria = id_categoria;
     }
